Replace the usage of is with proper comparation
pyflakes recently added a new kind of lint error that warns against the
use of `is` for checking str, bytes and int literals. The rationale
behind this new lint warning is described in
https://bugs.python.org/issue34850.

Similar discussions encourage treating is comparations as a bad practice
because of the uncertainty it might add during runtime.